services/python-ml/app/services/image_processor.py
"""Advanced image processing service with ML enhancement"""

import cv2
import numpy as np
from PIL import Image, ImageEnhance, ImageFilter
import torch
import torchvision.transforms as transforms
from typing import Dict, Any, Optional, Tuple
import asyncio
import aiohttp
import io
import base64
from loguru import logger

class ImageProcessor:
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.models = {}
        self._load_models()
        
    def _load_models(self):
        """Load pre-trained models for image processing"""
        try:
            # Load style transfer model
            self.models['style_transfer'] = torch.hub.load('pytorch/vision', 'vgg19', pretrained=True)
            self.models['style_transfer'].to(self.device)
            
            # Load super resolution model  
            self.models['super_resolution'] = torch.hub.load('ultralytics/yolov5', 'yolov5s', pretrained=True)
            
            logger.info("Image processing models loaded successfully")
        except Exception as e:
            logger.error(f"Failed to load models: {str(e)}")
    
    async def process_advanced(self, image_url: str, processing_type: str, parameters: Dict[str, Any]) -> Dict[str, Any]:
        """Main processing function with ML enhancement"""
        try:
            # Download image
            image = await self._download_image(image_url)
            
            # Apply processing based on type
            if processing_type == "style_transfer":
                result_image = await self._style_transfer(image, parameters)
            elif processing_type == "background_removal":
                result_image = await self._remove_background(image, parameters)
            elif processing_type == "color_enhancement":
                result_image = await self._enhance_colors(image, parameters)
            elif processing_type == "super_resolution":
                result_image = await self._super_resolution(image, parameters)
            elif processing_type == "fabric_analysis":
                return await self._analyze_fabric(image, parameters)
            elif processing_type == "design_optimization":
                result_image = await self._optimize_design(image, parameters)
            else:
                raise ValueError(f"Unknown processing type: {processing_type}")
            
            # Convert result to base64
            result_url = self._image_to_base64(result_image)
            
            return {
                "success": True,
                "result_url": result_url,
                "metadata": {
                    "processing_type": processing_type,
                    "original_size": image.size,
                    "result_size": result_image.size,
                    "parameters_used": parameters
                },
                "processing_time": 0.0  # Will be calculated by caller
            }
            
        except Exception as e:
            logger.error(f"Image processing failed: {str(e)}")
            return {
                "success": False,
                "error": str(e),
                "metadata": {"processing_type": processing_type}
            }
    
    async def _download_image(self, image_url: str) -> Image.Image:
        """Download image from URL"""
        async with aiohttp.ClientSession() as session:
            async with session.get(image_url) as response:
                if response.status == 200:
                    image_data = await response.read()
                    return Image.open(io.BytesIO(image_data))
                else:
                    raise Exception(f"Failed to download image: {response.status}")
    
    async def _style_transfer(self, image: Image.Image, parameters: Dict[str, Any]) -> Image.Image:
        """Apply neural style transfer"""
        try:
            style_strength = parameters.get("style_strength", 0.5)
            target_size = parameters.get("target_size", (512, 512))
            
            # Resize image
            image = image.resize(target_size)
            
            # Convert to tensor
            transform = transforms.Compose([
                transforms.ToTensor(),
                transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
            ])
            
            input_tensor = transform(image).unsqueeze(0).to(self.device)
            
            # Apply style transfer (simplified implementation)
            with torch.no_grad():
                # This is a placeholder - in real implementation, use proper style transfer model
                enhanced_tensor = input_tensor * (1 + style_strength * 0.2)
                enhanced_tensor = torch.clamp(enhanced_tensor, 0, 1)
            
            # Convert back to PIL Image
            result_array = enhanced_tensor.squeeze().cpu().numpy().transpose(1, 2, 0)
            result_array = (result_array * 255).astype(np.uint8)
            
            return Image.fromarray(result_array)
            
        except Exception as e:
            logger.error(f"Style transfer failed: {str(e)}")
            return image
    
    async def _remove_background(self, image: Image.Image, parameters: Dict[str, Any]) -> Image.Image:
        """Remove background using ML segmentation"""
        try:
            # Convert to OpenCV format
            cv_image = cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
            
            # Create mask using GrabCut algorithm
            mask = np.zeros(cv_image.shape[:2], np.uint8)
            bgd_model = np.zeros((1, 65), np.float64)
            fgd_model = np.zeros((1, 65), np.float64)
            
            # Define rectangle around object (simplified)
            height, width = cv_image.shape[:2]
            rect = (int(width*0.1), int(height*0.1), int(width*0.8), int(height*0.8))
            
            cv2.grabCut(cv_image, mask, rect, bgd_model, fgd_model, 5, cv2.GC_INIT_WITH_RECT)
            
            # Create final mask
            mask2 = np.where((mask == 2) | (mask == 0), 0, 1).astype('uint8')
            
            # Apply mask
            result = cv_image * mask2[:, :, np.newaxis]
            
            # Convert back to PIL
            result_rgb = cv2.cvtColor(result, cv2.COLOR_BGR2RGB)
            return Image.fromarray(result_rgb)
            
        except Exception as e:
            logger.error(f"Background removal failed: {str(e)}")
            return image
    
    async def _enhance_colors(self, image: Image.Image, parameters: Dict[str, Any]) -> Image.Image:
        """Enhanced color processing"""
        try:
            brightness = parameters.get("brightness", 1.0)
            contrast = parameters.get("contrast", 1.0)
            saturation = parameters.get("saturation", 1.0)
            
            # Apply enhancements
            if brightness != 1.0:
                enhancer = ImageEnhance.Brightness(image)
                image = enhancer.enhance(brightness)
            
            if contrast != 1.0:
                enhancer = ImageEnhance.Contrast(image)
                image = enhancer.enhance(contrast)
                
            if saturation != 1.0:
                enhancer = ImageEnhance.Color(image)
                image = enhancer.enhance(saturation)
            
            return image
            
        except Exception as e:
            logger.error(f"Color enhancement failed: {str(e)}")
            return image
    
    async def _super_resolution(self, image: Image.Image, parameters: Dict[str, Any]) -> Image.Image:
        """AI-powered super resolution"""
        try:
            scale_factor = parameters.get("scale_factor", 2)
            
            # Simple upscaling with interpolation (placeholder for real SR model)
            width, height = image.size
            new_size = (width * scale_factor, height * scale_factor)
            
            return image.resize(new_size, Image.Resampling.LANCZOS)
            
        except Exception as e:
            logger.error(f"Super resolution failed: {str(e)}")
            return image
    
    async def _analyze_fabric(self, image: Image.Image, parameters: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze fabric properties using computer vision"""
        try:
            # Convert to OpenCV format
            cv_image = cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
            
            # Analyze texture
            gray = cv2.cvtColor(cv_image, cv2.COLOR_BGR2GRAY)
            
            # Calculate texture features
            texture_variance = np.var(gray)
            texture_std = np.std(gray)
            
            # Analyze colors
            colors = cv_image.reshape(-1, 3)
            dominant_colors = self._get_dominant_colors(colors, k=5)
            
            # Fabric type classification (simplified)
            fabric_type = self._classify_fabric_type(texture_variance, dominant_colors)
            
            return {
                "success": True,
                "analysis": {
                    "fabric_type": fabric_type,
                    "texture_variance": float(texture_variance),
                    "dominant_colors": dominant_colors.tolist(),
                    "estimated_material": self._estimate_material(texture_variance),
                    "quality_score": min(100, max(0, 100 - texture_variance / 10))
                }
            }
            
        except Exception as e:
            logger.error(f"Fabric analysis failed: {str(e)}")
            return {"success": False, "error": str(e)}
    
    async def _optimize_design(self, image: Image.Image, parameters: Dict[str, Any]) -> Image.Image:
        """Optimize design using AI recommendations"""
        try:
            # Apply various optimizations
            optimization_type = parameters.get("type", "general")
            
            if optimization_type == "contrast":
                enhancer = ImageEnhance.Contrast(image)
                image = enhancer.enhance(1.2)
            elif optimization_type == "sharpness":
                image = image.filter(ImageFilter.UnsharpMask(radius=1, percent=150, threshold=3))
            elif optimization_type == "color_balance":
                enhancer = ImageEnhance.Color(image)
                image = enhancer.enhance(1.1)
            
            return image
            
        except Exception as e:
            logger.error(f"Design optimization failed: {str(e)}")
            return image
    
    def _get_dominant_colors(self, colors: np.ndarray, k: int = 5) -> np.ndarray:
        """Extract dominant colors using k-means clustering"""
        from sklearn.cluster import KMeans
        
        kmeans = KMeans(n_clusters=k, random_state=42, n_init=10)
        kmeans.fit(colors)
        return kmeans.cluster_centers_.astype(int)
    
    def _classify_fabric_type(self, texture_variance: float, dominant_colors: np.ndarray) -> str:
        """Simple fabric type classification"""
        if texture_variance > 1000:
            return "textured" 
        elif texture_variance > 500:
            return "medium_texture"
        else:
            return "smooth"
    
    def _estimate_material(self, texture_variance: float) -> str:
        """Estimate material type based on texture"""
        if texture_variance > 2000:
            return "wool"
        elif texture_variance > 1000:
            return "cotton"
        elif texture_variance > 500:
            return "linen"
        else:
            return "silk"
    
    def _image_to_base64(self, image: Image.Image) -> str:
        """Convert PIL Image to base64 string"""
        buffer = io.BytesIO()
        image.save(buffer, format='PNG')
        image_binary = buffer.getvalue()
        return base64.b64encode(image_binary).decode('utf-8')