services/python-ml/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
pillow==10.1.0
transformers==4.36.0
openai==1.3.8
anthropic==0.8.1
supabase==2.0.2
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0